fix: the same lock key and refactor duple
