+++ src/server/utils/data/index.ts
@@ -14,8 +14,9 @@ export * from "./get-volunteer-clones";
 export * from "./get-volunteer-form-data";
 export * from "./get-volunteer-patch-data";
 export * from "./get-volunteer-where";
-export * from "./sync-comment-tags";
 export * from "./is-trusted-domain";
+export * from "./run-w-advisory-lock";
+export * from "./sync-comment-tags";
 export * from "./update-leads";
 export * from "./write-agent-registration";
 export * from "./write-opportunity-contact-comment";
